Start using a forwarder pattern, using widgets as a use-case
