+++ app/src/main/java/fr/neamar/kiss/forwarder/Forwarder.java
@@ -1,0 +1,42 @@
+package fr.neamar.kiss.forwarder;
+
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.preference.PreferenceManager;
+import android.view.ContextMenu;
+import android.view.MenuItem;
+import android.view.View;
+
+import fr.neamar.kiss.MainActivity;
+
+public class Forwarder {
+    MainActivity mainActivity;
+    protected SharedPreferences prefs;
+
+    Forwarder(MainActivity mainActivity) {
+        this.mainActivity = mainActivity;
+        prefs = PreferenceManager.getDefaultSharedPreferences(mainActivity);
+    }
+
+    public void onCreate() {
+    }
+
+    public void onStart() {
+    }
+
+    public void onStop() {
+    }
+
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+    }
+
+    public boolean onOptionsItemSelected(MenuItem item) {
+        return false;
+    }
+
+    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
+    }
+
+    public void onDataSetChanged() {
+    }
+}
