+++ app/src/main/java/fr/neamar/kiss/forwarder/WidgetForwarder.java
@@ -1,0 +1,235 @@
+package fr.neamar.kiss.forwarder;
+
+import android.app.Activity;
+import android.appwidget.AppWidgetHost;
+import android.appwidget.AppWidgetHostView;
+import android.appwidget.AppWidgetManager;
+import android.appwidget.AppWidgetProviderInfo;
+import android.content.Context;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.os.Build;
+import android.view.ContextMenu;
+import android.view.MenuItem;
+import android.view.View;
+import android.view.ViewGroup;
+
+import java.util.HashMap;
+
+import fr.neamar.kiss.MainActivity;
+import fr.neamar.kiss.R;
+
+public class WidgetForwarder extends Forwarder {
+    private static final int REQUEST_PICK_APPWIDGET = 9;
+    private static final int REQUEST_CREATE_APPWIDGET = 5;
+
+    private static final int APPWIDGET_HOST_ID = 442;
+    private static final String WIDGET_PREFERENCE_ID = "fr.neamar.kiss.widgetprefs";
+
+    private SharedPreferences widgetPrefs;
+
+    /**
+     * Widget fields
+     */
+    private AppWidgetManager mAppWidgetManager;
+    private AppWidgetHost mAppWidgetHost;
+    private boolean widgetUsed = false;
+
+    /**
+     * View widgets are added to
+     */
+    private ViewGroup widgetArea;
+
+    WidgetForwarder(MainActivity mainActivity) {
+        super(mainActivity);
+    }
+
+    @Override
+    public void onCreate() {
+        // Initialize widget manager and host, restore widgets
+        widgetPrefs = mainActivity.getSharedPreferences(WIDGET_PREFERENCE_ID, Context.MODE_PRIVATE);
+
+        mAppWidgetManager = AppWidgetManager.getInstance(mainActivity);
+        mAppWidgetHost = new AppWidgetHost(mainActivity, APPWIDGET_HOST_ID);
+        widgetArea = (ViewGroup) mainActivity.findViewById(R.id.widgetLayout);
+
+        restoreWidgets();
+    }
+
+    @Override
+    public void onStart() {
+        // Start listening for widget update
+        mAppWidgetHost.startListening();
+    }
+
+    @Override
+    public void onStop() {
+        // Stop listening for widget update
+        mAppWidgetHost.stopListening();
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (resultCode == Activity.RESULT_OK) {
+            switch (requestCode) {
+                case REQUEST_CREATE_APPWIDGET:
+                    addAppWidget(data);
+                    break;
+                case REQUEST_PICK_APPWIDGET:
+                    configureAppWidget(data);
+                    break;
+            }
+        } else if (resultCode == Activity.RESULT_CANCELED && data != null) {
+            //if widget was not selected, delete id
+            int appWidgetId = data.getIntExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, -1);
+            if (appWidgetId != -1) {
+                mAppWidgetHost.deleteAppWidgetId(appWidgetId);
+            }
+        }
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if(item.getItemId() == R.id.widget) {
+            if (!widgetUsed) {
+                // request widget picker, a selection will lead to a call of onActivityResult
+                int appWidgetId = mAppWidgetHost.allocateAppWidgetId();
+                Intent pickIntent = new Intent(AppWidgetManager.ACTION_APPWIDGET_PICK);
+                pickIntent.putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, appWidgetId);
+                mainActivity.startActivityForResult(pickIntent, REQUEST_PICK_APPWIDGET);
+            } else {
+                // if we already have a widget we remove it
+                removeAllWidgets();
+            }
+            return true;
+        }
+
+        return false;
+    }
+
+    @Override
+    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
+        if (prefs.getBoolean("history-hide", true)) {
+            if (widgetUsed) {
+                menu.findItem(R.id.widget).setTitle(R.string.menu_widget_remove);
+            } else {
+                menu.findItem(R.id.widget).setTitle(R.string.menu_widget_add);
+            }
+        } else {
+            menu.findItem(R.id.widget).setVisible(false);
+        }
+    }
+
+    @Override
+    public void onDataSetChanged() {
+        if (!widgetUsed && mainActivity.adapter.isEmpty()) {
+            // when a widget is displayed this would prevent touches on the widget
+            mainActivity.emptyListView.setVisibility(View.VISIBLE);
+        }
+    }
+
+    /**
+     * Restores all previously added widgets
+     */
+    private void restoreWidgets() {
+        HashMap<String, Integer> widgetIds = (HashMap<String, Integer>) widgetPrefs.getAll();
+        for (int appWidgetId : widgetIds.values()) {
+            addWidgetToLauncher(appWidgetId);
+        }
+    }
+
+    /**
+     * Adds a widget to the widget area on the MainActivity
+     *
+     * @param appWidgetId id of widget to add
+     */
+    private void addWidgetToLauncher(int appWidgetId) {
+        // only add widgets if in minimal mode (may need launcher restart when turned on)
+        if (prefs.getBoolean("history-hide", true)) {
+            // remove empty list view when using widgets, this would block touches on the widget
+            mainActivity.emptyListView.setVisibility(View.GONE);
+            //add widget to view
+            AppWidgetProviderInfo appWidgetInfo = mAppWidgetManager.getAppWidgetInfo(appWidgetId);
+            if (appWidgetInfo == null) {
+                removeAllWidgets();
+                return;
+            }
+            AppWidgetHostView hostView = mAppWidgetHost.createView(mainActivity, appWidgetId, appWidgetInfo);
+            hostView.setMinimumHeight(appWidgetInfo.minHeight);
+            hostView.setAppWidget(appWidgetId, appWidgetInfo);
+            if (Build.VERSION.SDK_INT > 15) {
+                hostView.updateAppWidgetSize(null, appWidgetInfo.minWidth, appWidgetInfo.minHeight, appWidgetInfo.minWidth, appWidgetInfo.minHeight);
+            }
+            widgetArea.addView(hostView);
+        }
+        // only one widget allowed so widgetUsed is true now, even if not added to view
+        widgetUsed = true;
+    }
+
+    /**
+     * Removes all widgets from the launcher
+     */
+    private void removeAllWidgets() {
+        while (widgetArea.getChildCount() > 0) {
+            AppWidgetHostView widget = (AppWidgetHostView) widgetArea.getChildAt(0);
+            removeAppWidget(widget);
+        }
+    }
+
+    /**
+     * Removes a single widget and deletes it from persistent prefs
+     *
+     * @param hostView instance of a displayed widget
+     */
+    private void removeAppWidget(AppWidgetHostView hostView) {
+        // remove widget from view
+        int appWidgetId = hostView.getAppWidgetId();
+        mAppWidgetHost.deleteAppWidgetId(appWidgetId);
+        widgetArea.removeView(hostView);
+        // remove widget id from persistent prefs
+        SharedPreferences.Editor widgetPrefsEditor = widgetPrefs.edit();
+        widgetPrefsEditor.remove(String.valueOf(appWidgetId));
+        widgetPrefsEditor.apply();
+        // only one widget allowed so widgetUsed is false now
+        widgetUsed = false;
+    }
+
+    /**
+     * Adds widget to Activity and persists it in prefs to be able to restore it
+     *
+     * @param data Intent holding widget id to add
+     */
+    private void addAppWidget(Intent data) {
+        int appWidgetId = data.getIntExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, -1);
+        //add widget
+        addWidgetToLauncher(appWidgetId);
+        // Save widget in preferences
+        SharedPreferences.Editor widgetPrefsEditor = widgetPrefs.edit();
+        widgetPrefsEditor.putInt(String.valueOf(appWidgetId), appWidgetId);
+        widgetPrefsEditor.apply();
+    }
+
+    /**
+     * Check if widget needs configuration and display configuration view if necessary,
+     * otherwise just add the widget
+     *
+     * @param data Intent holding widget id to configure
+     */
+    private void configureAppWidget(Intent data) {
+        int appWidgetId = data.getIntExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, -1);
+
+        AppWidgetProviderInfo appWidget =
+                mAppWidgetManager.getAppWidgetInfo(appWidgetId);
+
+        if (appWidget.configure != null) {
+            // Launch over to configure widget, if needed.
+            Intent intent = new Intent(AppWidgetManager.ACTION_APPWIDGET_CONFIGURE);
+            intent.setComponent(appWidget.configure);
+            intent.putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, appWidgetId);
+            mainActivity.startActivityForResult(intent, REQUEST_CREATE_APPWIDGET);
+        } else {
+            // Otherwise, finish adding the widget.
+            addAppWidget(data);
+        }
+    }
+}
